New device fixture factories. (#57)
This PR separates out the new device fixture factories and their tests, so that they don't have to be in PR #56 anymore.

While doing that, it also changes how the fixtures handle attribute values, to make it match how the real Hubitat does it. They are no longer stored in device state, but in a separate map of attribute values that can be queried with the device's currentValue() method.